basic auth

+++ src/main/java/com/back/dao/AdresseRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.back.entities.Adresse; 
 
-public interface AdresseRepository  extends JpaRepository<Adresse, Integer>{
+public interface AdresseRepository  extends JpaRepository<Adresse, Long>{
 
 }
