+++ src/main/java/com/back/dao/AgenceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.back.entities.Agence; 
 
-public interface AgenceRepository  extends JpaRepository<Agence, Integer>{
+public interface AgenceRepository  extends JpaRepository<Agence, Long>{
 
 }
