+++ src/main/java/com/back/dao/CompteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.back.entities.Compte;
 
-public interface CompteRepository  extends JpaRepository<Compte, Integer>{
+public interface CompteRepository  extends JpaRepository<Compte, Long>{
 
 }
